+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- support for conversion from java.util.Date to LocalDate/LocalDateTime
 - fj-bom version set to 1.6.5
 - DBUtils.indentifyDB() now recognizes h2 (600) and hsqldb (700) databases.
 - IdGenerator for h2 is mapped to Postgres by default